+++ packages/espressocash_app/lib/app.dart
@@ -8,9 +8,11 @@ import 'core/analytics/analytics_manager.dart';
 import 'di.dart';
 import 'features/accounts/services/accounts_bloc.dart';
 import 'features/app_lock/module.dart';
+import 'features/authenticated/screens/authenticated_flow_screen.dart';
+import 'features/sign_in/screens/sign_in_flow_screen.dart';
 import 'l10n/localizely_updater.dart';
 import 'routes.dart';
-import 'routes.gr.dart';
+import 'ui/splash_screen.dart';
 import 'ui/theme.dart';
 
 class CryptopleaseApp extends StatefulWidget {
@@ -45,11 +47,11 @@ class _CryptopleaseAppState extends State<CryptopleaseApp> {
             _router,
             routes: (_) => [
               if (isAuthenticated)
-                const AuthenticatedFlowRoute()
+                AuthenticatedFlowScreen.route()
               else if (isLoading)
-                const SplashRoute()
+                SplashScreen.route()
               else
-                const SignInFlowRoute(),
+                SignInFlowScreen.route(),
             ],
             navigatorObservers: () => [
               sl<AnalyticsManager>().analyticsObserver,