+++ packages/espressocash_app/lib/features/accounts/services/accounts_state.dart
@@ -4,7 +4,6 @@ part of 'accounts_bloc.dart';
 class AccountsState with _$AccountsState {
   const factory AccountsState({
     MyAccount? account,
-    @Default(false) bool hasFinishedOnboarding,
     @Default(false) bool isProcessing,
   }) = _AccountsState;
 }