+++ packages/espressocash_app/lib/features/investments/widgets/crypto_investments.dart
@@ -1,5 +1,4 @@
 import 'package:decimal/decimal.dart';
-import 'package:dfunc/dfunc.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -7,60 +6,58 @@ import 'package:sliver_tools/sliver_tools.dart';
 
 import '../../../core/amount.dart';
 import '../../../core/presentation/format_amount.dart';
+import '../../../core/presentation/value_stream_builder.dart';
 import '../../../core/tokens/token.dart';
-import '../../../core/user_preferences.dart';
+import '../../../di.dart';
 import '../../../l10n/device_locale.dart';
 import '../../../l10n/l10n.dart';
 import '../../../ui/colors.dart';
 import '../../../ui/theme.dart';
-import '../../balances/services/balances_bloc.dart';
-import '../../balances/widgets/watch_balance.dart';
+import '../../conversion_rates/services/watch_user_total_fiat_balance.dart';
 import '../data/repository.dart';
+import '../services/watch_investments.dart';
 import 'portfolio_widget.dart';
 
 class CryptoInvestments extends StatelessWidget {
   const CryptoInvestments({super.key});
 
-  static final Decimal _minimumUsdAmount = Decimal.parse('0.01');
+  @override
+  Widget build(BuildContext context) => ValueStreamBuilder<Amount>(
+        create: () => sl<WatchUserTotalFiatBalance>().call(
+          ignoreTokens: [Token.usdc],
+        ),
+        builder: (context, balance) {
+          final displayEmptyBalances = context
+              .watch<InvestmentSettingsRepository>()
+              .displayEmptyBalances;
+
+          return balance.decimal == Decimal.zero && !displayEmptyBalances
+              ? const SliverToBoxAdapter(child: SizedBox.shrink())
+              : MultiSliver(
+                  children: [
+                    _Header(balance),
+                    const SizedBox(height: 15),
+                    _Portfolio(displayEmptyBalances: displayEmptyBalances),
+                  ],
+                );
+        },
+      );
+}
+
+class _Portfolio extends StatelessWidget {
+  const _Portfolio({required this.displayEmptyBalances});
+
+  final bool displayEmptyBalances;
 
   @override
-  Widget build(BuildContext context) {
-    final displayEmptyBalances =
-        context.watch<InvestmentSettingsRepository>().displayEmptyBalances;
-
-    final balance = context.watchUserTotalFiatBalance(
-      context.watch<UserPreferences>().fiatCurrency,
-      ignoreTokens: [Token.usdc],
-    );
-
-    return balance.decimal == Decimal.zero && !displayEmptyBalances
-        ? const SliverToBoxAdapter(child: SizedBox.shrink())
-        : MultiSliver(
-            children: [
-              _Header(balance),
-              const SizedBox(height: 15),
-              BlocBuilder<BalancesBloc, BalancesState>(
-                builder: (context, state) {
-                  final tokens =
-                      state.userTokens.where((e) => e != Token.usdc).let(
-                            (tokens) => displayEmptyBalances
-                                ? tokens
-                                : tokens.where((token) {
-                                    final Decimal balance = context
-                                            .watchUserFiatBalance(token)
-                                            ?.decimal ??
-                                        Decimal.zero;
-
-                                    return balance >= _minimumUsdAmount;
-                                  }),
-                          );
-
-                  return PortfolioWidget(tokens: IList(tokens));
-                },
-              ),
-            ],
-          );
-  }
+  Widget build(BuildContext context) => ValueStreamBuilder<IList<Token>>(
+        create: () => (
+          sl<WatchInvestments>()
+              .call(displayEmptyBalances: displayEmptyBalances),
+          const IListConst([]),
+        ),
+        builder: (context, tokens) => PortfolioWidget(tokens: tokens),
+      );
 }
 
 class _Header extends StatelessWidget {