+++ packages/espressocash_app/lib/features/onboarding/screens/no_email_and_password_screen.dart
@@ -2,16 +2,20 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
 import '../../../l10n/l10n.dart';
+import '../../../routes.gr.dart';
 import '../../../ui/app_bar.dart';
 import '../../../ui/back_button.dart';
 import '../../../ui/info_widget.dart';
 import '../../../ui/onboarding_screen.dart';
 import '../../../ui/theme.dart';
-import 'onboarding_flow_screen.dart';
 
 @RoutePage()
 class NoEmailAndPasswordScreen extends StatelessWidget {
-  const NoEmailAndPasswordScreen({super.key});
+  const NoEmailAndPasswordScreen({super.key, required this.onDone});
+
+  static const route = NoEmailAndPasswordRoute.new;
+
+  final VoidCallback onDone;
 
   @override
   Widget build(BuildContext context) => CpTheme.dark(
@@ -19,8 +23,7 @@ class NoEmailAndPasswordScreen extends StatelessWidget {
           body: OnboardingScreen(
             footer: OnboardingFooterButton(
               text: context.l10n.iUnderstand,
-              onPressed: () => context.onboardingRouter
-                  .onExplainNoEmailAndPasswordCompleted(),
+              onPressed: onDone,
             ),
             children: [
               CpAppBar(