+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/services/oskp_service.dart
@@ -128,10 +128,8 @@ class OSKPService {
       switch (apiVersion) {
         case SplitKeyApiVersion.manual:
           response = await _client.createPayment(dto);
-          break;
         case SplitKeyApiVersion.smartContract:
           response = await _client.createPaymentEc(dto);
-          break;
       }
       final tx = await response.transaction
           .let(SignedTx.decode)
@@ -177,7 +175,6 @@ class OSKPService {
           tx = await transaction
               .let(SignedTx.decode)
               .let((it) => it.resign(LocalWallet(escrow)));
-          break;
         case SplitKeyApiVersion.smartContract:
           final response = await _client.cancelPaymentEc(dto);
           transaction = response.transaction;
@@ -185,7 +182,6 @@ class OSKPService {
           tx = await transaction
               .let(SignedTx.decode)
               .let((it) => it.resign(account));
-          break;
       }
 
       return OSKPStatus.cancelTxCreated(