+++ packages/espressocash_app/lib/features/outgoing_split_key_payments/services/tx_ready_watcher.dart
@@ -46,10 +46,8 @@ class TxReadyWatcher {
           switch (payment.apiVersion) {
             case SplitKeyApiVersion.manual:
               destinationAccounts = tx.getDestinations();
-              break;
             case SplitKeyApiVersion.smartContract:
               destinationAccounts = txDetails.getInnerDestinations();
-              break;
           }
 
           final newStatus = await destinationAccounts.let(