+++ packages/espressocash_app/lib/features/payment_request/screens/link_request_flow_screen.dart
@@ -1,11 +1,10 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
 
 import '../../../core/amount.dart';
-import '../../../core/user_preferences.dart';
+import '../../../core/flow.dart';
 import '../../../di.dart';
 import '../../../routes.gr.dart';
 import '../../../ui/dialogs.dart';
@@ -15,6 +14,7 @@ import '../../conversion_rates/data/repository.dart';
 import '../data/repository.dart';
 import '../services/create_payment_request_bloc.dart';
 import '../widgets/request_note_screen.dart';
+import 'link_details_flow_screen.dart';
 
 @RoutePage()
 class LinkRequestFlowScreen extends StatefulWidget {
@@ -22,6 +22,8 @@ class LinkRequestFlowScreen extends StatefulWidget {
     super.key,
     required this.initialAmount,
   });
+
+  static const route = LinkRequestFlowRoute.new;
 
   final CryptoAmount initialAmount;
 
@@ -38,7 +40,6 @@ class _LinkRequestFlowScreenState extends State<LinkRequestFlowScreen> {
     final amount = widget.initialAmount;
 
     _paymentRequestBloc = CreatePaymentRequestBloc(
-      userCurrency: context.read<UserPreferences>().fiatCurrency,
       repository: sl<PaymentRequestRepository>(),
       conversionRatesRepository: sl<ConversionRatesRepository>(),
     );
@@ -96,20 +97,16 @@ class _ContentState extends State<_Content> implements NoteSetter {
         child:
             BlocConsumer<CreatePaymentRequestBloc, CreatePaymentRequestState>(
           listenWhen: (s1, s2) => s1.flow != s2.flow,
-          listener: (context, state) => state.flow.maybeWhen(
-            failure: (error) => showErrorDialog(
-              context,
-              'Failed to send money',
-              error,
-            ),
-            success: (request) {
-              context.router.popUntilRoot();
-              context.navigateTo(LinkDetailsFlowRoute(id: request.id));
-            },
-            orElse: ignore,
-          ),
+          listener: (context, state) => switch (state.flow) {
+            FlowFailure(:final error) =>
+              showErrorDialog(context, 'Failed to send money', error),
+            FlowSuccess(:final result) => context
+              ..router.popUntilRoot()
+              ..navigateTo(LinkDetailsFlowScreen.route(id: result.id)),
+            _ => null,
+          },
           builder: (context, state) => CpLoader(
-            isLoading: state.flow.isProcessing(),
+            isLoading: state.flow.isProcessing,
             child: RequestNoteScreen(amount: widget.amount),
           ),
           buildWhen: (s1, s2) => s1.flow != s2.flow,