+++ packages/espressocash_app/lib/features/payment_request/screens/share_payment_request_screen.dart
@@ -4,6 +4,7 @@ import 'package:provider/provider.dart';
 
 import '../../../di.dart';
 import '../../../l10n/l10n.dart';
+import '../../../routes.gr.dart';
 import '../../../ui/app_bar.dart';
 import '../../../ui/back_button.dart';
 import '../../../ui/dialogs.dart';
@@ -18,6 +19,8 @@ import '../widgets/share_qr_code.dart';
 @RoutePage()
 class SharePaymentRequestScreen extends StatelessWidget {
   const SharePaymentRequestScreen({super.key});
+
+  static const route = SharePaymentRequestRoute.new;
 
   @override
   Widget build(BuildContext context) {
@@ -64,9 +67,10 @@ class SharePaymentRequestScreen extends StatelessWidget {
                   text: context.l10n.paymentRequestCancel,
                   onPressed: () => showConfirmationDialog(
                     context,
-                    title: context.l10n.cancelPaymentRequestConfirmationTitle,
-                    message:
-                        context.l10n.cancelPaymentRequestConfirmationSubtitle,
+                    title:
+                        context.l10n.paymentRequest_lblCancelConfirmationTitle,
+                    message: context
+                        .l10n.paymentRequest_lblCancelConfirmationSubtitle,
                     onConfirm: () {
                       sl<PaymentRequestRepository>().delete(request.id);
                       context.router.pop();